moved kafka publish to common
